// src/app/models/user.model.ts
export class User {
    id: string;
    firstName: string;
    lastName: string;
    dob: Date;
    email: string;
    regNo: string;
    password: string;
    isAdmin: boolean;
    class: string;
    section: string;
    phone: string;
    city: string;
    isActive: boolean;
    resetToken: string;
}